feature контроль прав на удаление

+++ controllers/cardsControllers.js
@@ -1,31 +1,31 @@
-import { constants as http2Constants } from 'node:http2';
-import mongoose from 'mongoose';
-import Card from '../models/cardModel.js';
+import { constants as http2Constants } from "node:http2";
+import mongoose from "mongoose";
+import Card from "../models/cardModel.js";
 
 function errorHandler(error, res) {
   if (error instanceof mongoose.Error.ValidationError) {
     return res.status(http2Constants.HTTP_STATUS_BAD_REQUEST).send({
       message: `${Object.values(error.errors)
         .map((err) => err.message)
-        .join(', ')}`,
+        .join(", ")}`,
     });
   }
 
   if (error instanceof mongoose.Error.DocumentNotFoundError) {
     return res.status(http2Constants.HTTP_STATUS_NOT_FOUND).send({
-      message: 'Запрашиваемая карточка не найдена',
+      message: "Запрашиваемая карточка не найдена",
     });
   }
 
   if (error instanceof mongoose.Error.CastError) {
     return res.status(http2Constants.HTTP_STATUS_BAD_REQUEST).send({
-      message: 'Некорректный id',
+      message: "Некорректный id",
     });
   }
 
   return res
     .status(http2Constants.HTTP_STATUS_INTERNAL_SERVER_ERROR)
-    .send({ message: 'Server Error' });
+    .send({ message: "Server Error" });
 }
 
 export const getAllCards = async (req, res) => {
@@ -50,10 +50,13 @@ export const addNewCard = async (req, res) => {
 
 export const deleteCardById = async (req, res) => {
   try {
+    const card = await Card.findById(req.params.cardId).orFail();
+    console.log(card.owner, req.user._id);
+    if (card.owner != req.user._id) throw new Error("Недостаточно прав"); //! Сделать адекватно
     await Card.findByIdAndDelete(req.params.cardId).orFail();
     res
       .status(http2Constants.HTTP_STATUS_OK)
-      .send({ message: 'Успешно удалено!' });
+      .send({ message: "Успешно удалено!" });
   } catch (error) {
     errorHandler(error, res);
   }
@@ -64,7 +67,7 @@ export const likeCard = async (req, res) => {
     const card = await Card.findByIdAndUpdate(
       req.params.cardId,
       { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
-      { new: true },
+      { new: true }
     ).orFail();
 
     res.status(http2Constants.HTTP_STATUS_CREATED).send(card);
@@ -78,7 +81,7 @@ export const dislikeCard = async (req, res) => {
     const card = await Card.findByIdAndUpdate(
       req.params.cardId,
       { $pull: { likes: req.user._id } }, // убрать _id из массива
-      { new: true },
+      { new: true }
     ).orFail();
     res.status(http2Constants.HTTP_STATUS_OK).send(card);
   } catch (error) {
